Fixes the Final boss crash
Fixes the Final boss from crashing when adding a new function to PlayerObject.txt. Waht is going on here is that the integer values being set here are actually the values of functions.  What happens is when you add a new function in the PlayerObject, the function values are incremented by one. So, in the EggMobile.txt script where Object.State=59 is no longer the value 59 but now 60. To fix this, all that was needed was to change the values to point to their function call instead (which in the end, becomes a value that is NOT hard coded, but assigned during compile time).
The BossCurtain.txt object has an initialize call into the EggMobile.txt object and needed to be changed as well.
NOTE: there are still other objects within the game that break when adding functions.