Fixed a couple of other script function errors
Options Menu was calling the value of `6`.  The comment was oddly named implying that it should call `OptionsMenuC_Function6` but that doesn't seem to be the case. Calling `OptionsMenu_Function6 ` seems fine.
Moving Block is interesting. Rotation isn't actually used for... rotation. Its a function pointer basically. This seems to somehow fix the ice bug.